backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
drf-yasg==1.21.7
django-sslserver==0.22
django-ratelimit==4.1.0
Pillow==10.1.0
python-decouple==3.8
psycopg2-binary==2.9.7
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
whitenoise==6.6.0
